+++ client/src/App.js
@@ -1,26 +1,15 @@
-import { useEffect } from "react";
-import axios from "axios";
+
 import logo from "./logo.svg";
 import "./App.css";
 
 function App() {
-  useEffect(() => {
-    axios
-      .get("/api/config")
-      .then((response) => {
-        console.log(response.data);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, []);
 
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/App.js</code> and save to reload.
+          Hello
         </p>
         <a
           className="App-link"
